Detect and use POSIX (and BSD if POSIX is not available) signals to protect the critical sections in the code.  This code was already present in cvs, but was never activated unless one put "-DPOSIX" into the Makefile or CFLAGS.  This change totally cures our excessive stray lock problem on FreeBSD (users are pressing ^C during cvs update).
This checks for the existance of sigaction and sigprocmask in
configure. If so, system.h defines POSIX_SIGNALS.  The various code
segments were changed to be #ifdef POSIX_SIGNALS rather than #ifdef POSIX
which was never set before.  Same for BSD_SIGNALS which is activated
if !POSIX_SIGNALS, and sigvec, sigblock and sigsetmask exist.

CVS-ID: lib/ChangeLog 1.27
CVS-ID: lib/sighandle.c 1.2
CVS-ID: lib/system.h 1.10
CVS-ID: src/ChangeLog 1.195
CVS-ID: src/run.c 1.3
CVS-ID: ChangeLog 1.83
CVS-ID: config.h.in 1.13
CVS-ID: configure 1.28
CVS-ID: configure.in 1.26