* client.c (update_entries): Fix memory leak; free mode_string and file_timestamp. (send_fileproc): Fix memory leak; call freevers_ts before exiting.
* module.c (do_module): Partially fix memory leak; added
variable so that the address of memory allocated by line2argv
is retained, but comment out the call to free_names.  Freeing
the vector at that point loses because some of the elements
may be used later in the function.
(cat_module): fix memory leak.

* recurse.c (start_recursion): Fix memory leak; free return
value of Name_Repository after it has been used.

CVS-ID: src/ChangeLog 1.222
CVS-ID: src/client.c 1.30