* Makefile.in (Makefile): cause this target to fail after rebuilding Makefile to force you to run make again to use the new Makefile.  [this can be a bit of a pain but it's the only portable way to ensure the user doesn't build the wrong configuration]
* Makefile.in (configure, config.h.in, stamp.h.in): re-enable
these targets, but make them just print a message about *not*
running autoconf and/or autoheader.  Also change the comment to
note that 'make dist' should really be the mechanism used for
commenting out the real rules and adding the echo statements.
(stamp-h): stamp-h and config.h must both be built in tandem by
running config.status to keep the dependencies in order.
(config.status): add setting of CC for the recheck (ala CFLAGS).

CVS-ID: Makefile.in 1.55