[These mods make "checkout" work with "-d /absolute/pathname" once again.]
* checkout.c (checkout_proc): the -d flag on the command line
should override the -d flag in the modules file if the latter is
an absolute path.  The loop that assembles the list of directories
to build has been reorganized slightly to prepare for rewriting
with last_component rather than assuming '/' as a path separator.
Also added to that loop was some code to handle absolute
pathnames.
(build_dirs_and_chdir): add a new argument that tells this routine
whether or not to check before it creates and populates
directories or not.

* sanity.sh (toplevel-12): change test to reflect the new style of
this error message.

* create_adm.c (Create_Admin): include the directory in the error
message.

* filesubr.c (last_component): return the top-level directory when
asked about the top-level directory. [Same mod in all directories.]

CVS-ID: emx/ChangeLog 1.22
CVS-ID: emx/filesubr.c 1.4
CVS-ID: macintosh/ChangeLog 1.48
CVS-ID: macintosh/filesubr.c 1.5
CVS-ID: os2/ChangeLog 1.179
CVS-ID: os2/filesubr.c 1.29
CVS-ID: src/ChangeLog 1.1443
CVS-ID: src/checkout.c 1.77
CVS-ID: src/create_adm.c 1.30
CVS-ID: src/filesubr.c 1.37
CVS-ID: src/sanity.sh 1.361
CVS-ID: vms/ChangeLog 1.49
CVS-ID: vms/filesubr.c 1.8
CVS-ID: windows-NT/ChangeLog 1.116
CVS-ID: windows-NT/filesubr.c 1.31