* sanity.sh (?): Additional test for ampersand module behavior with '-d'.
    for Greg Klanderman  <greg@itasoftware.com>

* checkout.c (build_one_dir): Fix typo where clauses of two
conditionals were reversed in call to Create_Admin.  This caused
the CVS/Tag file to be removed in cases where it should have been
set, and vice-versa.  It only surfaced in rare cases as this code
is only invoked when using the -d option to build the path to
check out in.  Further, the bug would only matter when checking
out a module containing ampersand modules within it, via
client/server CVS.

CVS-ID: src/ChangeLog 1.2067
CVS-ID: src/checkout.c 1.96
CVS-ID: src/sanity.sh 1.658